requirements.txt
asgiref==3.8.1
bcrypt==3.2.2
cffi==1.17.1
diskcache==5.6.3
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-yasg==1.21.10
inflection==0.5.1
Jinja2==3.1.6
llama_cpp_python==0.3.9
MarkupSafe==3.0.2
mongoengine==0.29.1
numpy==2.2.6
packaging==25.0
passlib==1.7.4
pycparser==2.22
PyJWT==2.9.0
pymongo==4.13.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
